// ammerli/apps/backend/src/api/request/enums/request-status.enum.ts
export enum RequestStatusEnum {
  PENDING = 'PENDING',
  SEARCHING = 'SEARCHING',
  DISPATCHED = 'DISPATCHED',
  LOCKED = 'LOCKED',
  ACCEPTED = 'ACCEPTED',
  ARRIVED = 'ARRIVED',
  DELIVERING = 'DELIVERING',
  DELIVERED = 'DELIVERED',
  CANCELLED = 'CANCELLED',
  EXPIRED = 'EXPIRED',
  SCHEDULED = 'SCHEDULED',
}